+++ api/server/router/swarm/helpers.go
@@ -94,5 +94,19 @@ func adjustForAPIVersion(cliVersion string, service *swarm.ServiceSpec) {
 			// API version 1.40
 			service.TaskTemplate.Placement.MaxReplicas = 0
 		}
+
+		if service.TaskTemplate.Resources != nil {
+			if service.TaskTemplate.Resources.SwapBytes != nil {
+				// Setting services' swap size wasn't supported before
+				// API version 1.40
+				service.TaskTemplate.Resources.SwapBytes = nil
+			}
+
+			if service.TaskTemplate.Resources.MemorySwappiness != nil {
+				// Setting services' swappiness wasn't supported before
+				// API version 1.40
+				service.TaskTemplate.Resources.MemorySwappiness = nil
+			}
+		}
 	}
 }